+++ R/valid.R
@@ -1,21 +1,42 @@
 #' Generic validation function
 #'
-#' @param choice
-#' @param choices
-#' @param reverse
-#' @param flip
-#' @param strict
+#' Designed to be called *inside* a custom`valid_*` function (e.g. [valid_yes_no()])
+#'
+#' @param choice ([character]) Available values
+#' @param choices ([character]) Selection of available values
+#' @param reverse ([logical]) Reverse order yes/no
+#' @param flip ([logical]) Flip names and values yes/no
+#' @param strict ([logical]) Be strict about "things that might go wrong" yes/no
+#' @param unname ([logical]) Drop names yes/no
 #'
 #' @return
 #' @export
 #'
 #' @examples
+#' valid_foo <- function(
+#'   choice = character(),
+#'   ...
+#' ) {
+#'   choices <- letters[1:3]
+#'
+#'   names(choices) <- choices %>% toupper()
+#'
+#'   valid::valid(
+#'     choice = choice,
+#'     choices = choices,
+#'     ...
+#'   )
+#' }
+#' valid_foo()
+#' valid_foo("a")
+#' try(valid_foo("d"))
 valid <- function(
   choice = character(),
   choices = character(),
   reverse = FALSE,
   flip = FALSE,
-  strict = TRUE
+  strict = TRUE,
+  unname = FALSE
 ) {
   out <- if (length(choice)) {
     # if (choice %in% names(choices)) {
@@ -33,6 +54,14 @@ valid <- function(
     choices
   }
 
+  # Reverse
+  out <- if (!reverse) {
+    out
+  } else {
+    rev(out)
+  }
+
+  # Flip
   out <- if (!flip) {
     out
   } else {
@@ -40,20 +69,25 @@ valid <- function(
     out <- flip_values_and_names(out)
   }
 
-  out <- if (!reverse) {
-    out
-  } else {
-    rev(out)
-  }
 
+  # Strict yes/no
   if (!length(out) && strict) {
     call <- as.character(sys.call(-1)[[1]])
     stop("Invalid choice: {call}({deparse(choice)})" %>%
         stringr::str_glue())
   }
 
+  # Unname
+  out <- if (unname) {
+    out %>% unname()
+  } else {
+    out
+  }
+
   out
 }
+
+# Helpers ------------------------------------------------------------------
 
 #' Flip values and names
 #'
@@ -66,8 +100,6 @@ flip_values_and_names <- function(x) {
   names(x) <- names
   x
 }
-
-# Helpers ------------------------------------------------------------------
 
 #' Is answer really TRUE or FALSE?
 #'
@@ -99,69 +131,105 @@ not_in <- function(x, set) {
 
 #' Valid: yes/no
 #'
-#' @param choice [character] Actual choice out of all available choices
-#' @param flip ([logical]) Flip the vector's order?
+#' @param choice ([character]) Actual choice out of all available choices
+#' @param ... Further arguments that will be passed to [valid::valid()]
 #'
 #' @return
 #' @export
-valid_yes_no <- function(choice = character(), flip = FALSE) {
+#'
+#' @examples
+#' valid_yes_no()
+#' valid_yes_no("yes")
+#' try(valid_yes_no("invalid"))
+#' valid_yes_no(reverse = TRUE)
+#' valid_yes_no(flip = TRUE)
+#' valid_yes_no(unname = TRUE)
+valid_yes_no <- function(
+  choice = character(),
+  ...
+) {
   valid::valid(
     choice = choice,
     choices = c(
       yes = "Yes",
       no = "No"
     ),
-    flip = flip
+    ...
   )
 }
 
 #' Valid: again/exit
 #'
-#' @param choice [character] Actual choice out of all available choices
-#' @param flip ([logical]) Flip the vector's order?
+#' @param choice ([character]) Selection from available valid choices
+#' @param ... Further arguments that will be passed to [valid::valid()])]
 #'
 #' @return
 #' @export
-valid_again_exit <- function(choice = character(), flip = FALSE) {
+#' @examples
+#' valid_again_exit()
+#' valid_again_exit("again")
+#' try(valid_again_exit("invalid"))
+#' valid_again_exit(reverse = TRUE)
+#' valid_again_exit(flip = TRUE)
+#' valid_again_exit(unname = TRUE)
+valid_again_exit <- function(
+  choice = character(),
+  ...
+) {
   valid::valid(
     choice = choice,
     choices = c(
       again = "Let me start over",
       exit = "Exit"
     ),
-    flip = flip
+    ...
   )
 }
 
 #' Valid: none
 #'
-#' @param choice [character] Actual choice out of all available choices
-#' @param flip ([logical]) Flip the vector's order?
+#' @param choice ([character]) Selection from available valid choices
+#' @param ... Further arguments that should be passed to [valid::valid()]
 #'
 #' @return
 #' @export
-valid_none <- function(choice = character(), flip = FALSE) {
+#' @examples
+#' valid_none()
+#' valid_none("none")
+#' try(valid_none("invalid"))
+#' valid_none(reverse = TRUE)
+#' valid_none(flip = TRUE)
+#' valid_none(unname = TRUE)
+valid_none <- function(
+  choice = character(),
+  ...
+) {
   valid::valid(
     choice = choice,
     choices = c(
       none = "None"
     ),
-    flip = flip
+    ...
   )
 }
 
 #' Valid: yes/no/again/exit
 #'
-#' @param choice [character] Actual choice out of all available choices
-#' @param flip ([logical]) Flip names and values?
-#' @param reverse ([logical]) Flip order?
+#' @param choice ([character]) Selection from available valid choices
+#' @param ... Further arguments that will be passed to [valid::valid()]
 #'
 #' @return
 #' @export
+#' @examples
+#' valid_yes_no_again_exit()
+#' valid_yes_no_again_exit("yes")
+#' try(valid_yes_no_again_exit("invalid"))
+#' valid_yes_no_again_exit(reverse = TRUE)
+#' valid_yes_no_again_exit(flip = TRUE)
+#' valid_yes_no_again_exit(unname = TRUE)
 valid_yes_no_again_exit <- function(
   choice = character(),
-  flip = FALSE,
-  reverse = FALSE
+  ...
 ) {
   valid::valid(
     choice = choice,
@@ -169,8 +237,7 @@ valid_yes_no_again_exit <- function(
       valid_yes_no(),
       valid_again_exit()
     ),
-    flip = flip,
-    reverse = reverse
+    ...
   )
 }
 
@@ -178,24 +245,28 @@ valid_yes_no_again_exit <- function(
 
 #' Valid authentication
 #'
-#' @param auth ([character]) Authentication choice
-#' @param flip ([logical]) Flip names and values?
-#' @param reverse ([logical00]) Reverse order?
+#' @param auth ([character]) Selection from available valid choices
+#' @param ... Further arguments that will be passed to [valid::valid()]
 #'
 #' @return
 #' @export
+#' @examples
+#' valid_authentication()
+#' valid_authentication("ssh")
+#' try(valid_authentication("invalid"))
+#' valid_authentication(reverse = TRUE)
+#' valid_authentication(flip = TRUE)
+#' valid_authentication(unname = TRUE)
 valid_authentication <- function(
   auth = character(),
-  flip = FALSE,
-  reverse = FALSE
+  ...
 ) {
   auths <- c("ssh", "https")
   names(auths) <- auths
   valid::valid(
     choice = auth,
     choices = auths,
-    flip = flip,
-    reverse = reverse
+    ...
   )
 }
 
@@ -203,20 +274,21 @@ valid_authentication <- function(
 
 #' Valid DevOps environments
 #'
-#' @param devops_env ([character]) DevOps environment choice
-#' @param flip ([logical]) Flip names and values?
-#' @param reverse ([logical]) Flip order?
+#' @param devops_env ([character]) Selection from available valid choices
+#' @param ... Further arguments that will be passed to [valid::valid()]
 #'
 #' @return
 #' @export
 #' @examples
 #' valid_devops_envs()
 #' valid_devops_envs("staging")
-#' valid_devops_envs("staging", reverse = TRUE)
+#' try(valid_devops_envs("invalid"))
+#' valid_devops_envs(reverse = TRUE)
+#' valid_devops_envs(flip = TRUE)
+#' valid_devops_envs(unname = TRUE)
 valid_devops_envs <- function(
   devops_env = character(),
-  flip = FALSE,
-  reverse = FALSE
+  ...
 ) {
   values <- c("dev", "staging", "prod")
   names(values) <- values
@@ -224,8 +296,7 @@ valid_devops_envs <- function(
   valid::valid(
     choice = devops_env,
     choices = values,
-    flip = flip,
-    reverse = reverse
+    ...
   )
 }
 
@@ -234,15 +305,19 @@ valid_devops_envs <- function(
 #' Valid licenses
 #'
 #' @param license [character] License choice
-#' @param flip ([logical]) Flip names and values?
-#' @param reverse ([logical]) Flip order?
-#'
+#' @param ... Further arguments that will be passed to [valid::valid()]#'
 #' @return
 #' @export
+#' @examples
+#' valid_licenses()
+#' valid_licenses("gpl3)
+#' try(valid_licenses("invalid"))
+#' valid_licenses(reverse = TRUE)
+#' valid_licenses(flip = TRUE)
+#' valid_licenses(unname = TRUE)
 valid_licenses <- function(
   license = character(),
-  flip = FALSE,
-  reverse = FALSE
+  ...
 ) {
   licenses <- c("GPL v3", "MIT", "CC0", "CCBY 4.0", "LGPL v3", "APL 2.0", "AGPL v3")
   names <- c("gpl3", "mit", "cc0", "ccby", "lgpl", "apl2", "agpl3")
@@ -250,8 +325,7 @@ valid_licenses <- function(
   valid::valid(
     choice = license,
     choices = licenses,
-    flip = flip,
-    reverse = reverse
+    ...
   )
 }
 
@@ -260,22 +334,26 @@ valid_licenses <- function(
 #' Valid dependency types
 #'
 #' @param type ([character]) Dependency type choice
-#' @param flip ([logical]) Flip names and values?
-#' @param reverse ([logical]) Flip order?
+#' @param ... Further arguments that will be passed to [valid::valid()]
 #'
 #' @return
 #' @export
+#' @examples
+#' valid_dep_types()
+#' valid_dep_types("Suggests")
+#' try(valid_dep_types("invalid"))
+#' valid_dep_types(reverse = TRUE)
+#' valid_dep_types(flip = TRUE)
+#' valid_dep_types(unname = TRUE)
 valid_dep_types <- function(
   type = character(),
-  flip = FALSE,
-  reverse = FALSE
+  ...
 ) {
   types <- c("Suggests", "Imports", "Depends", "Enhances", "LinkingTo")
   names(types) <- types
   valid::valid(
     choice = type,
     choices = types,
-    flip = flip,
-    reverse = reverse
+    ...
   )
 }