+++ tests/testthat/test-valids.R
@@ -1,14 +1,77 @@
+# Foo ---------------------------------------------------------------------
+
+valid_foo <- function(
+  choice = character(),
+  ...
+) {
+  choices <- letters[1:3]
+
+  names(choices) <- choices %>% toupper()
+
+  valid::valid(
+    choice = choice,
+    choices = choices,
+    ...
+  )
+}
+
+test_that("Foo", {
+  result <- valid_foo()
+  expected <- c(A = "a", B = "b", C = "c")
+  expect_identical(result, expected)
+})
+
+test_that("Reverse", {
+  result <- valid_foo(reverse = TRUE)
+  expected <- c(C = "c", B = "b", A = "a")
+  expect_identical(result, expected)
+})
+
+test_that("Flip", {
+  result <- valid_foo(flip = TRUE)
+  expected <- c(a = "A", b = "B", c = "C")
+  expect_identical(result, expected)
+})
+
+test_that("Unname", {
+  result <- valid_foo(unname = TRUE)
+  expected <- c("a", "b", "c")
+  expect_identical(result, expected)
+})
+
+# Helpers -----------------------------------------------------------------
+
+test_that("Is", {
+  result <- "yes" %>% is_answer_true_false()
+  expect_true(result)
+
+  result <- "no" %>% is_answer_true_false()
+  expect_true(result)
+
+  result <- "exit" %>% is_answer_true_false()
+  expect_false(result)
+
+  result <- "abc" %>% is_answer_true_false()
+  expect_false(result)
+})
+
+# Yes/no ------------------------------------------------------------------
+
 test_that("Yes/no", {
   result <- valid_yes_no()
   expected <- structure(c("Yes", "No"), names = c("yes", "no"))
   expect_identical(result, expected)
 })
 
+# Again/exit --------------------------------------------------------------
+
 test_that("Again/exit", {
   result <- valid_again_exit()
   expected <- c(again = "Let me start over", exit = "Exit")
   expect_identical(result, expected)
 })
+
+# None --------------------------------------------------------------------
 
 test_that("None", {
   result <- valid_none()
@@ -25,20 +88,6 @@ test_that("Yes/no/again/exit", {
     exit = "Exit"
   )
   expect_identical(result, expected)
-})
-
-test_that("Is", {
-  result <- "yes" %>% is_answer_true_false()
-  expect_true(result)
-
-  result <- "no" %>% is_answer_true_false()
-  expect_true(result)
-
-  result <- "exit" %>% is_answer_true_false()
-  expect_false(result)
-
-  result <- "abc" %>% is_answer_true_false()
-  expect_false(result)
 })
 
 # Authentication ----------------------------------------------------------
@@ -88,4 +137,3 @@ test_that("Package dependency types", {
     )
   expect_identical(result, expected)
 })
-